cast to avoid error
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@17201 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/drivers/tcpdg/src/nm_tcpdg.c
@@ -885,7 +885,7 @@ nm_tcpdg_send_iov	(struct nm_pkt_wrap *p_pw) {
                 }
 
                 do {
-                        uint64_t len	= tbx_min(p_cur->iov_len, ret);
+                        uint64_t len	= tbx_min(p_cur->iov_len, (uint64_t)ret);
                         p_cur->iov_base	+= len;
                         p_cur->iov_len	-= len;
